% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badges.R
\name{badge_document_status}
\alias{badge_document_status}
\title{Add a badge to a document on it's status.}
\usage{
badge_document_status(state = document_status$state)
}
\arguments{
\item{state}{What state the project is in (completed, in progress). See \code{document_status$state}}
}
\value{
Sends badge text to screen and puts it to the clipboard.
}
\description{
Add a badge to a document on it's status.
}
