% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensing.R
\name{ccby_icon_html}
\alias{ccby_icon_html}
\title{HTML code for inserting CCY-BY icon.}
\usage{
ccby_icon_html()
}
\value{
Character vector of HTML source for icon.
}
\description{
HTML code for inserting CCY-BY icon.
}
