% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft-rmd.R
\name{draft_bookdown_chapter}
\alias{draft_bookdown_chapter}
\title{Create a bookdown chapter file.}
\usage{
draft_bookdown_chapter(.filename)
}
\arguments{
\item{.filename}{The name of the file to create.}
}
\value{
Creates a bookdown chapter file.
}
\description{
Create a bookdown chapter file.
}
