% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logos.R
\name{logo}
\alias{logo}
\title{File path of logo.}
\usage{
logo(org = logo_list())
}
\arguments{
\item{org}{Organization name. See \code{\link[=logo_list]{logo_list()}} for list of names.}
}
\value{
A file path to the logo image.
}
\description{
File path of logo.
}
