% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwj-package.R
\docType{package}
\name{lwj-package}
\alias{lwj}
\alias{lwj-package}
\title{lwj: My Own Functions}
\description{
This contains functions I use in my daily work that don't yet have 
    their own package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lwjohnst86/lwj}
  \item Report bugs at \url{https://github.com/lwjohnst86/lwj/issues}
}

}
\author{
\strong{Maintainer}: Luke Johnston \email{lwjohnst@gmail.com} (\href{https://orcid.org/0000-0003-4169-2616}{ORCID})

}
\keyword{internal}
