% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{my_rstudio_config}
\alias{my_rstudio_config}
\title{Set basic configurations for using RStudio.}
\usage{
my_rstudio_config()
}
\value{
NULL. The function is used for its side effects.
}
\description{
Set basic configurations for using RStudio.
}
