% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcid.R
\name{orcid_icon_html}
\alias{orcid_icon_html}
\title{ORCID number icon for HTML.}
\usage{
orcid_icon_html()
}
\value{
Character vector of the HTML source for the icon.
}
\description{
ORCID number icon for HTML.
}
