% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr-code.R
\name{qr_to_svg}
\alias{qr_to_svg}
\title{QR Code generator to SVG file.}
\usage{
qr_to_svg(text, path)
}
\arguments{
\item{text}{The text or URL to use to add to the QR code.}

\item{path}{The output path for the SVG file.}
}
\description{
QR Code generator to SVG file.
}
